+++ oneflow/core/boxing/boxing_dividor_util.h
@@ -13,11 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_BOXING_DIVIDOR_UTIL_H_
-#define ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_BOXING_DIVIDOR_UTIL_H_
+#ifndef ONEFLOW_CORE_BOXING_BOXING_DIVIDOR_UTIL_H_
+#define ONEFLOW_CORE_BOXING_BOXING_DIVIDOR_UTIL_H_
 
 #include "oneflow/core/common/device_type.pb.h"
-#include "oneflow/core/framework/op_interpreter/boxing/boxing_dividor.h"
+#include "oneflow/core/boxing/boxing_dividor.h"
 
 namespace oneflow {
 
@@ -34,4 +34,4 @@ extern Maybe<BoxingDividor> (*OutFirstDeviceAndAllBroadcast)();
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_BOXING_DIVIDOR_UTIL_H_
+#endif  // ONEFLOW_CORE_BOXING_BOXING_DIVIDOR_UTIL_H_
