+++ oneflow/core/boxing/eager_boxing_interpreter.cpp
@@ -16,9 +16,9 @@ limitations under the License.
 #include <typeinfo>
 #include "oneflow/core/common/container_util.h"
 #include "oneflow/core/common/registry_error.h"
-#include "oneflow/core/framework/op_interpreter/boxing/eager_boxing_interpreter.h"
+#include "oneflow/core/boxing/eager_boxing_interpreter.h"
 #include "oneflow/core/framework/tensor_rpc_util.h"
-#include "oneflow/core/framework/op_interpreter/boxing/eager_boxing_interpreter_mgr.h"
+#include "oneflow/core/boxing/eager_boxing_interpreter_mgr.h"
 
 namespace oneflow {
