+++ oneflow/core/boxing/eager_boxing_interpreter.h
@@ -13,12 +13,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_EAGER_BOXING_INTERPRETER_H_
-#define ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_EAGER_BOXING_INTERPRETER_H_
+#ifndef ONEFLOW_CORE_BOXING_EAGER_BOXING_INTERPRETER_H_
+#define ONEFLOW_CORE_BOXING_EAGER_BOXING_INTERPRETER_H_
 
 #include "oneflow/core/common/symbol.h"
 #include "oneflow/core/framework/tensor_tuple.h"
-#include "oneflow/core/framework/op_interpreter/boxing/boxing_dividor.h"
+#include "oneflow/core/boxing/boxing_dividor.h"
 #include "oneflow/core/framework/tensor.h"
 #include "oneflow/core/framework/placed_nd_sbp.h"
 #include "oneflow/core/job/parallel_desc.h"
@@ -190,4 +190,4 @@ Maybe<BoxingExprIf> OptionalBoxing(const std::string& boxing_mame);
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_EAGER_BOXING_INTERPRETER_H_
+#endif  // ONEFLOW_CORE_BOXING_EAGER_BOXING_INTERPRETER_H_
