+++ oneflow/core/boxing/eager_boxing_interpreter_util.h
@@ -13,8 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_EAGER_BOXING_INTERPRETER_UTIL_H_
-#define ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_EAGER_BOXING_INTERPRETER_UTIL_H_
+#ifndef ONEFLOW_CORE_BOXING_EAGER_BOXING_INTERPRETER_UTIL_H_
+#define ONEFLOW_CORE_BOXING_EAGER_BOXING_INTERPRETER_UTIL_H_
 
 #include "oneflow/core/common/symbol.h"
 #include "oneflow/core/job/parallel_desc.h"
@@ -40,4 +40,4 @@ struct EagerBoxingInterpreterUtil {
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_FRAMEWORK_OP_INTERPRETER_BOXING_EAGER_BOXING_INTERPRETER_UTIL_H_
+#endif  // ONEFLOW_CORE_BOXING_EAGER_BOXING_INTERPRETER_UTIL_H_
