+++ oneflow/core/boxing/naive_b2p_boxing_interpreter.cpp
@@ -13,30 +13,34 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/control/global_process_ctx.h"
-#include "oneflow/core/framework/nd_sbp.h"
+#include "oneflow/core/boxing/eager_boxing_interpreter.h"
+#include "oneflow/core/boxing/eager_boxing_interpreter_util.h"
 #include "oneflow/core/framework/device.h"
-#include "oneflow/core/framework/op_interpreter/boxing/eager_boxing_interpreter_util.h"
-#include "oneflow/core/framework/op_interpreter/boxing/eager_boxing_interpreter.h"
+#include "oneflow/core/framework/nd_sbp.h"
+#include "oneflow/core/job/global_for.h"
+#include "oneflow/core/job/resource_desc.h"
+#include "oneflow/core/control/global_process_ctx.h"
 #include "oneflow/core/functional/functional.h"
-#include "oneflow/core/common/decorator.h"
 
 namespace oneflow {
 
 namespace {
 
-Maybe<void> RawCheckNaive1ToP(Symbol<PlacedNdSbp> in, Symbol<PlacedNdSbp> out) {
-  CHECK_EQ_OR_RETURN(in->placement()->parallel_num(), 1);
+Maybe<void> RawCheckNaiveBToP(Symbol<PlacedNdSbp> in, Symbol<PlacedNdSbp> out) {
+  CHECK_EQ_OR_RETURN(in->nd_sbp()->sbp_parallel_size(), 1);
+  CHECK_EQ_OR_RETURN(out->nd_sbp()->sbp_parallel_size(), 1);
+  CHECK_OR_RETURN(EagerBoxingInterpreterUtil::IsAllBroadcastNdSbp(in->nd_sbp()));
   CHECK_OR_RETURN(EagerBoxingInterpreterUtil::IsAllPartialSumNdSbp(out->nd_sbp()));
-  CHECK_OR_RETURN(out->placement()->Bigger(*in->placement()));
+
+  CHECK_OR_RETURN(in->placement() == out->placement());
   return Maybe<void>::Ok();
 }
 
-static constexpr auto* CheckNaive1ToP = DECORATE(&RawCheckNaive1ToP, ThreadLocal);
+static constexpr auto* CheckNaiveBToP = DECORATE(&RawCheckNaiveBToP, ThreadLocal);
 
 }  // namespace
 
-Maybe<one::Tensor> Naive1ToP(const std::shared_ptr<one::Tensor>& tensor, Symbol<PlacedNdSbp> in,
+Maybe<one::Tensor> NaiveBToP(const std::shared_ptr<one::Tensor>& tensor, Symbol<PlacedNdSbp> in,
                              Symbol<PlacedNdSbp> out) {
   const auto& tensor_nd_sbp = JUST(tensor->nd_sbp());
   CHECK_OR_RETURN(tensor_nd_sbp == in->nd_sbp());
@@ -45,19 +49,17 @@ Maybe<one::Tensor> Naive1ToP(const std::shared_ptr<one::Tensor>& tensor, Symbol<
 
   int64_t root = JUST(tensor_placement->MachineId4ParallelId(0));
   std::shared_ptr<one::Tensor> local_tensor = JUST(tensor->cur_rank_phy_tensor());
-  const auto& out_parallel_id = JUST(GetParallelId4CurrentProcessCtx(out->placement()));
-  if (root == GlobalProcessCtx::Rank() || !out_parallel_id->has_value()) {
+  if (root == GlobalProcessCtx::Rank()) {
     // do nothing
   } else {
     const std::string& device_type = Device::Type4DeviceTag(tensor_placement->device_tag());
-    local_tensor = JUST(one::functional::Constant(*tensor->shape(), 0, tensor->dtype(),
-                                                  JUST(Device::New(device_type))));
+    local_tensor = JUST(one::functional::ZerosLike(local_tensor));
   }
   return JUST(one::functional::LocalToConsistent(local_tensor, out->placement(),
                                                  *JUST(GetSbpList(out->nd_sbp())), *tensor->shape(),
                                                  tensor->dtype()));
 }
 
-COMMAND(RegisterBoxingFunction("naive-1-to-p", CheckNaive1ToP, &Naive1ToP));
+COMMAND(RegisterBoxingFunction("naive-b-to-p", CheckNaiveBToP, &NaiveBToP));
 
 }  // namespace oneflow
